Mood propagation working

+++ src/core/Components/AttributeComponent.hpp
@@ -62,11 +62,11 @@ namespace Core
 				/*! The rioter's conceptual group ID. Used for determining which flow field to use.*/
 				int groupID;
 
-				/*! The rioter's rage represented by an integer with range [0, 100]. */
-				int rage;
+				/*! The rioter's rage represented by an floating point value with range [0, 100]. */
+				float rage;
 
-				/*! The rioter's pressure represented by an integer with range [0, 100]. */
-				int pressure;
+				/*! The rioter's pressure represented by an floating point value with range [0, 100]. */
+				float pressure;
 
 				/*! The rioter's alignment represented by the RioterAlignment enum. */
 				RioterAlignment alignment;
